# docs/api-definitions.yml
openapi: "3.0.2"
info:
  title: Designful API
  version: "1.0.0"
servers:
  - url: https://api.designful.com/api/v1
    description: Production server
paths:
...
  /posts:
    post:
      summary: Create a new post
      security:
        - bearerAuth: []
      tags:
        - post
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                  example: "Shaver Cool Biometric"
                user_id:
                  type: string
                  example: "user-clvuk13be0001winzqsx3n8h5"
      responses:
        "200":
          description: Post created successfully.
        "400":
          description: Bad request if the request body is not correct.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Invalid request"
    get:
      summary: Get a list of posts
      tags:
        - post
      description: Retrieves a list of posts, filtered by user and paginated.
      parameters:
        - in: query
          name: userId
          required: false
          schema:
            type: string
          description: The user ID to filter posts by.
          example: "user-clvuk13be0001winzqsx3n8h5"
        - in: query
          name: limit
          required: false
          schema:
            type: integer
            default: 10
            minimum: 1
            maximum: 100
          description: The maximum number of posts to return at one time. Defaults to 10.
        - in: query
          name: offset
          required: false
          schema:
            type: integer
            default: 0
            minimum: 0
          description: The offset from which to start listing the posts.
      responses:
        "200":
          description: An array of posts.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Post"
        "400":
          description: Bad request parameters.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Failed to parse pagination limit or offset"
        "500":
          description: Internal server error, such as database errors.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Failed to fetch posts."
  /posts/{id}:
    get:
      summary: Get a post by ID
      tags:
        - post
      description: Retrieves detailed information about a specific post including associated user details, categories, total likes, and whether the current user has liked the post.
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
          description: The unique identifier of the post.
          example: "post-clvwcht6m002y6aubd4cy9nhn"
      responses:
        "200":
          description: Detailed information about the post.
          content:
            application/json:
              schema:
                type: object
                properties:
                  post:
                    $ref: "#/components/schemas/Post"
                  user:
                    $ref: "#/components/schemas/User"
                  likes:
                    type: integer
                    description: The total number of likes this post has received.
                  userLiked:
                    type: boolean
                    description: A boolean flag indicating whether the current user has liked the post.
                  userPurchased:
                    type: boolean
                    description: A boolean flag indicating whether the current user has purchased the post.
        "404":
          description: Post not found.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: object
                    example: "The post not found"
        "500":
          description: Internal server error, such as database errors.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Failed to fetch post data."

    patch:
      summary: Update a post
      security:
        - bearerAuth: []
      tags:
        - post
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                  example: "Shaver Cool Biometric"
                description:
                  type: string
                  example: "Our hers many somebody hourly fortnightly as indeed bale how of normally upon aha shall everyone anybody mine with homeless some Gaussian those cautiously which me swallow of Shakespearean problem Sudanese one pout of sew currency way alas less enough."
                price:
                  type: integer
                  example: 200
                is_featured:
                  type: boolean
                status:
                  type: boolean
      responses:
        "200":
          description: Post updated successfully.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Post"
        "400":
          description: Invalid input provided.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Invalid request data."

    delete:
      summary: Delete a post
      security:
        - bearerAuth: []
      tags:
        - post
      responses:
        "200":
          description: Post deleted successfully.
        "404":
          description: Post not found.
...
  /categories/{id}/posts:
    get:
      summary: Get posts by category ID
      tags:
        - category
      description: Retrieves posts associated with a specific category.
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            description: The unique identifier of the category.
      responses:
        "200":
          description: A list of posts in the specified category.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Post"
        "404":
          description: Category not found.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "The category not found."
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    User:
      type: object
      properties:
        id:
          type: string
          description: "The unique identifier of the user."
          example: user-clvwcht6m002y6aubd4cy9nhn
          readOnly: true
        name:
          type: string
          description: "The name of the user."
          example: Michael Jackson
        username:
          type: string
          description: "The username of the user, automatically assigned a CUID if not provided."
          example: "johndoe2024"
        profile_image_url:
          type: string
          nullable: true
          description: "Optional URL to the user's profile image."
        stripe_account_id:
          type: string
          nullable: true
          description: "Optional Stripe customer ID for the user."
          example: acct_1Nv0FGQ9RKHgCVdK
        url:
          type: string
          nullable: true
          example: https://www.example.com
          description: "Optional URL associated with the user."
        email:
          type: string
          description: "The user's email address."
          example: "jackson@example.com"
        email_verified:
          type: boolean
          default: false
          description: "Flag indicating whether the user's email address has been verified."
        bio:
          type: string
          nullable: true
          description: "A brief biography of the user."
          example: Explorer of virtual vistas and real-world mountains alike, I spend my days weaving through codes and nights under stars. An ardent lover of coffee and conversations, I believe every cat has a secret, and every book hides a universe. When not in front of a screen, you'll find me sketching out my next big adventure or lost in the melodies of forgotten songs. Dreamer, doer, and occasional philosopher, chasing light, capturing moments, and coding dreams into reality
        dob:
          type: string
          nullable: true
          description: "Date of birth of the user."
        role:
          type: string
          enum: [fan, creator, admin]
          description: "The role of the user within the system."
          default: fan
        created_at:
          type: string
          format: date-time
          default: "Current Timestamp"
          description: "Timestamp when the user account was created."
        updated_at:
          type: string
          format: date-time
          default: "Current Timestamp"
          description: "Timestamp when the user account was last updated, automatically updated to current timestamp on modification."
...
    Post:
      type: object
      required:
        - id
        - title
        - created_at
        - updated_at
      properties:
        id:
          type: string
          description: "Unique identifier for the post."
          example: post-clvwcht6m002y6aubd4cy9nhn
          readOnly: true
        title:
          type: string
          description: "Title of the post."
        description:
          type: string
          nullable: true
          description: "Detailed description of the post. Optional."
        thumbnail_url:
          type: string
          description: "URL of the thumbnail image for the post. Optional."
        video_url:
          type: string
          description: "URL of the video. Optional."
        mux_asset_id:
          type: string
          nullable: true
          description: "Mux Asset ID that can be used to reference videos within Mux. Optional."
        mux_playback_id:
          type: string
          nullable: true
          description: "Mux Playback ID used to play the video through Mux's CDN. Optional."
        price:
          type: integer
          nullable: true
          description: "Price of the video in cents. A value of 0 or null (if allowed) indicates the video is free."
        is_featured:
          type: boolean
          default: false
          description: "Indicates if the post is featured. Defaults to false. Nillable."
        status:
          type: boolean
          default: false
          description: "Status of the post, typically true for active. Defaults to false. Nillable."
        created_at:
          type: string
          format: date-time
          description: "The date and time when the post was created. Defaults to the current time."
        updated_at:
          type: string
          format: date-time
          description: "The date and time when the post was last updated. Automatically updates when the post is modified."
...