# src/api/ent/predicate/predicate.go
// Code generated by ent, DO NOT EDIT.

package predicate

import (
	"entgo.io/ent/dialect/sql"
)

// Category is the predicate function for category builders.
type Category func(*sql.Selector)

// Like is the predicate function for like builders.
type Like func(*sql.Selector)

// Order is the predicate function for order builders.
type Order func(*sql.Selector)

// Post is the predicate function for post builders.
type Post func(*sql.Selector)

// Subscription is the predicate function for subscription builders.
type Subscription func(*sql.Selector)

// User is the predicate function for user builders.
type User func(*sql.Selector)

// VerificationToken is the predicate function for verificationtoken builders.
type VerificationToken func(*sql.Selector)